requirements.txt
langdetect
jieba
python-pinyin
pyxdg
tabulate[widechars]
summa
stopwordsiso
